Make HTTPHeader::contentLength return an off_t, and use String.toOffset() to retrieve its value (large file support)
git-svn-id: https://svn.code.sf.net/p/dansguardian/code/trunk@816 c6e6bffc-5bc6-4360-8735-ace3d95093af

+++ dansguardian/src/HTTPHeader.hpp
@@ -73,7 +73,7 @@ public:
 	String requestType();
 	int returnCode();
 	// get content length - returns -1 if undetermined
-	int contentLength();
+	off_t contentLength();
 	String getContentType();
 	// check received content type against given content type
 	bool isContentType(const String& t);
@@ -155,7 +155,7 @@ private:
 	std::string cachedurl;
 
 	char postdata[15];
-	int postdatalen;
+	off_t postdatalen;
 	bool postdatachopped;
 	bool ispostupload;
